Switch to density-based levelling, no accounting for owned range
Density here is defined simply as the size divided by
token range. If there are holes in the locally owned
range, the density will be underestimated.
Shards are no longer necessary to determine a UCS-
created SSTable's level.

+++ src/java/org/apache/cassandra/db/compaction/CompactionAggregate.java
@@ -716,7 +716,7 @@ public abstract class CompactionAggregate
             return bucket.index;
         }
 
-        public long bucketMin()
+        public double bucketMin()
         {
             return bucket.min;
         }
